% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowline_spawning_attr}
\alias{flowline_spawning_attr}
\title{NHD Flowline (ComID) Spawning Suitability Flags}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 178868 rows and 6 columns.
}
\usage{
flowline_spawning_attr
}
\description{
NHD Flowline (ComID) Spawning Suitability Flags
}
\examples{
flowline_spawning_attr |> pillar::glimpse()
}
\keyword{datasets}
