% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{glimpse_plot}
\alias{glimpse_plot}
\title{Glimpse plot}
\usage{
glimpse_plot(x, ...)
}
\arguments{
\item{x}{object of any class with valid \code{plot} method}

\item{...}{additional parameters passed to \code{plot} method}
}
\value{
the same object, invisibly, while printing a plot to the console
}
\description{
Glimpse plot
}
