% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_drc}
\alias{habitat_drc}
\title{Create Duration Suitability Rating Curve for a streamgage}
\usage{
habitat_drc(
  streamgage,
  habitat_type = "rearing",
  run = "fall",
  wy_group = "Dry",
  gradient = "Valley Foothill"
)
}
\arguments{
\item{streamgage}{Three-letter CDEC identifier for the streamgage}

\item{habitat_type}{Either "rearing" (default) or "spawning"}

\item{run}{One of "fall" (default), "late fall", "winter", "spring", or "steelhead"}

\item{wy_group}{Either "Dry" or "Wet" for water year type scenario}

\item{gradient}{For rearing, either "Valley Foothill" (default) or "Valley Lowland"}
}
\value{
A \code{tibble} with columns for flow (\code{flow_cfs}) and duration habitat suitability index (\code{durhsi})
}
\description{
Create Duration Suitability Rating Curve for a streamgage
}
\examples{

habitat_drc("FSB", habitat_type="rearing", gradient = "Valley Foothill")

}
