% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_drc_weighted}
\alias{habitat_drc_weighted}
\title{Create Duration Suitability Rating Curve for a streamgage (scaled to prediction point by drainage area and precipitation)}
\usage{
habitat_drc_weighted(
  streamgage,
  reach,
  scale = "durhsi",
  habitat_type = "rearing",
  run = "fall",
  wy_group = "Dry",
  gradient = "Valley Foothill"
)
}
\arguments{
\item{streamgage}{Three-letter CDEC identifier for the streamgage}

\item{reach}{Integer \code{comid} identifier for reach identifying the prediction point (used for scaling)}

\item{scale}{Either "durhsi" to keep original flow scale and adjust duration HSI values (default), or "flow" to keep original duration HSI scale and adjust flow values.}

\item{habitat_type}{Either "rearing" (default) or "spawning"}

\item{run}{One of "fall" (default), "late fall", "winter", "spring", or "steelhead"}

\item{wy_group}{Either "Dry" or "Wet" for water year type scenario}

\item{gradient}{For rearing, either "Valley Foothill" (default) or "Valley Lowland"}
}
\value{
A \code{tibble} with columns for flow (\code{flow_cfs}) and duration habitat suitability index (\code{durhsi})
}
\description{
Create Duration Suitability Rating Curve for a streamgage (scaled to prediction point by drainage area and precipitation)
}
\examples{

habitat_drc_weighted("FSB", habitat_type="rearing", gradient = "Valley Foothill", reach = 7978069)

}
