% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_fsa}
\alias{habitat_fsa}
\title{Create Flow-To-Suitable-Area Curve for Reach, Mainstem, or Watershed}
\usage{
habitat_fsa(reach, mainstem, watershed, habitat_type = "rearing", units = "ft")
}
\arguments{
\item{reach}{Integer \code{comid} identifier for reach (must provide either reach, mainstem, OR watershed)}

\item{mainstem}{String name for mainstem (refer to \code{cv_mainstems} dataset for options)}

\item{watershed}{String name for watershed (refer to \code{cv_watersheds} dataset for options)}

\item{habitat_type}{Either "rearing" (default) or "spawning"}

\item{units}{Either "ft" for square ft per linear ft, or "ac" for total acres}
}
\value{
A \code{tibble} with columns for flow (\code{flow_cfs}) and suitable habitat (\code{habitat})
}
\description{
Create Flow-To-Suitable-Area Curve for Reach, Mainstem, or Watershed
}
\examples{

habitat_fsa(reach = 7978069, units = "ft")

habitat_fsa(mainstem = "Feather River", units = "ac")

}
