% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_fsa_duration}
\alias{habitat_fsa_duration}
\title{Create Flow-To-Suitable-Area Curve (with Duration HSI applied) for Reach, Mainstem, or Watershed}
\usage{
habitat_fsa_duration(
  reach,
  mainstem,
  watershed,
  streamgage,
  habitat_type = "rearing",
  units = "ft",
  run = "fall",
  wy_group = "Dry",
  gradient = "Valley Foothill"
)
}
\arguments{
\item{reach}{Integer \code{comid} identifier for reach (must provide either reach, mainstem, OR watershed)}

\item{mainstem}{String name for mainstem (refer to \code{cv_mainstems} dataset for options)}

\item{watershed}{String name for watershed (refer to \code{cv_watersheds} dataset for options)}

\item{streamgage}{Three-letter CDEC identifier for the streamgage}

\item{habitat_type}{Either "rearing" (default) or "spawning"}

\item{units}{Desired units for output, either "ft" for square ft per linear ft, or "ac" for total acres}

\item{run}{One of "fall" (default), "late fall", "winter", "spring", or "steelhead"}

\item{wy_group}{Either "Dry" or "Wet" for water year type scenario}

\item{gradient}{For rearing, either "Valley Foothill" (default) or "Valley Lowland"}
}
\value{
A \code{tibble} with columns for flow (\code{flow_cfs}) and suitable habitat (\code{habitat})
}
\description{
Create Flow-To-Suitable-Area Curve (with Duration HSI applied) for Reach, Mainstem, or Watershed
}
\examples{

habitat_fsa_duration(reach = 7978069, streamgage = "FSB")

habitat_fsa_duration(mainstem = "Feather River", streamgage = "FSB")

}
