% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_fsa_duration_reach}
\alias{habitat_fsa_duration_reach}
\title{Create Flow-To-Suitable-Area Curve (with Duration HSI applied) for Reach}
\usage{
habitat_fsa_duration_reach(
  reach,
  streamgage,
  habitat_type = "rearing",
  units = "ft",
  run = "fall",
  wy_group = "Dry",
  gradient = "Valley Foothill"
)
}
\arguments{
\item{reach}{Integer \code{comid} identifier for reach}

\item{streamgage}{Three-letter CDEC identifier for the streamgage}

\item{habitat_type}{Either "rearing" or "spawning"}

\item{units}{Desired units for output, either "ft" for square ft per linear ft, or "ac" for total acres}

\item{run}{One of "fall" (default), "late fall", "winter", "spring", or "steelhead"}

\item{wy_group}{Either "Dry" or "Wet" for water year type scenario}

\item{gradient}{For rearing, either "Valley Foothill" (default) or "Valley Lowland"}
}
\value{
A \code{tibble} with columns for flow (\code{flow_cfs}) and suitable habitat (\code{habitat})
}
\description{
Create Flow-To-Suitable-Area Curve (with Duration HSI applied) for Reach
}
\examples{

habitat_fsa_duration_reach(reach = 7978069, streamgage = "FSB")

}
