% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_fsa_reach}
\alias{habitat_fsa_reach}
\title{Create Flow-To-Suitable-Area-Curve for Individual Reach (unscaled)}
\usage{
habitat_fsa_reach(reach, habitat_type = "rearing", units = "ft")
}
\arguments{
\item{reach}{Integer \code{comid} identifier for reach}

\item{habitat_type}{Either "rearing" or "spawning"}

\item{units}{Desired units for output, either "ft" for square ft per linear ft, or "ac" for total acres}
}
\value{
A \code{tibble} with columns for flow (\code{flow_cfs}) and suitable habitat (\code{habitat})
}
\description{
Create Flow-To-Suitable-Area-Curve for Individual Reach (unscaled)
}
\examples{

habitat_fsa_reach(reach = 7978069)

}
