% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_fsa_reach_scaled}
\alias{habitat_fsa_reach_scaled}
\title{Create Flow-To-Suitable-Area Curve for Individual Reach (scaled by multiplier)}
\usage{
habitat_fsa_reach_scaled(
  reach,
  multiplier,
  habitat_type = "rearing",
  units = "ft"
)
}
\arguments{
\item{reach}{Integer \code{comid} identifier for reach}

\item{multiplier}{Numeric multiplier to scale flow}

\item{habitat_type}{Either "rearing" (default) or "spawning"}

\item{units}{Desired units for output, either "ft" for square ft per linear ft, or "ac" for total acres}
}
\value{
A \code{tibble} with columns for flow (\code{flow_cfs}) and suitable habitat (\code{habitat})
}
\description{
Create Flow-To-Suitable-Area Curve for Individual Reach (scaled by multiplier)
}
\examples{

habitat_fsa_reach_scaled(reach = 7978069, multiplier = 0.8)

}
