% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_predict}
\alias{habitat_predict}
\title{Predict Habitat Area using a Flow-to-Suitable-Area Curve}
\usage{
habitat_predict(
  x,
  reach,
  mainstem,
  watershed,
  streamgage,
  fsa_x,
  fsa_y,
  habitat_type = "rearing",
  units = "ft",
  run = "fall",
  wy_group = "Dry",
  gradient = "Valley Foothill"
)
}
\arguments{
\item{x}{A vector of flows at which to predict suitable habitat values}

\item{reach}{Integer \code{comid} identifier for reach (must provide either reach, mainstem, OR watershed)}

\item{mainstem}{String name for mainstem (refer to \code{cv_mainstems} dataset for options)}

\item{watershed}{String name for watershed (refer to \code{cv_watersheds} dataset for options)}

\item{streamgage}{(optional) Three-letter CDEC identifier for a streamgage}

\item{fsa_x}{(optional) Vector of flows for manually defined flow-to-suitable-area curve. If provided, overrides reach, mainstem, or watershed lookup.}

\item{fsa_y}{(optional) Vector of habitat values for manually defined flow-to-suitable-area curve, of same length as \code{fsa_x}}

\item{habitat_type}{Either "rearing" (default) or "spawning"}

\item{units}{Desired units for output, either "ft" for square ft per linear ft, or "ac" for total acres}

\item{run}{(required if streamgage is provided) One of "fall" (default), "late fall", "winter", "spring", or "steelhead"}

\item{wy_group}{(required if streamgage is provided) Either "Dry" or "Wet" for water year type scenario}

\item{gradient}{(required if streamgage is provided) For rearing, either "Valley Foothill" (default) or "Valley Lowland"}
}
\value{
A vector of the same length as \code{x} containing predicted suitable habitat values
}
\description{
Predict Habitat Area using a Flow-to-Suitable-Area Curve
}
\examples{
# example non-temporal prediction given comid
habitat_predict(c(0, 10, 150, 300, 9999), reach = 7978069)

# example temporal prediction given comid and streamgage
habitat_predict(c(0, 10, 150, 300, 9999), reach = 7978069, streamgage = "FSB")

# example non-temporal prediction given mainstem
habitat_predict(c(0, 10, 150, 300, 9999), mainstem = "Feather River")

# example temporal prediction given comid and streamgage
habitat_predict(c(0, 10, 150, 300, 9999), mainstem = "Feather River", streamgage = "FSB")

# example manual prediction given custom provided flow-to-suitable-area curve
habitat_predict(c(0, 10, 150, 300, 9999), fsa_x = c(0, 100, 1000), fsa_y = c(47, 470, 4700))

}
