% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wua_predicted_cv_watersheds}
\alias{wua_predicted_cv_watersheds}
\title{WUA/LF by CV Watershed by Flow (predicted by statistical model)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2950 rows and 10 columns.
}
\usage{
wua_predicted_cv_watersheds
}
\description{
WUA/LF by CV Watershed by Flow (predicted by statistical model)
}
\examples{
wua_predicted_cv_watersheds |> pillar::glimpse()
}
\keyword{datasets}
